refactoring (to make more portable) some reflection stuff

+++ icecapvm/src/natives_allOS.c
@@ -860,7 +860,7 @@ int16 n_java_lang_Object_getClass(int32 *sp) {
 #if defined(N_JAVA_LANG_CLASS_GETSUPERCLASS)
 int16 n_java_lang_Class_getSuperclass(int32 *sp) {
 	Object* class = (Object*) (pointer) sp[0];
-	uint16 classIndex = ((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
+	uint16 classIndex = (uint16)((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
 	sp[0] = (int32) (pointer) getClass(classes[classIndex].superClass);
 	return -1;
 }
@@ -979,7 +979,7 @@ Object* getClass(unsigned short classIndex)
 {
 	Object *class = head;
 	while (class != 0) {
-		if (((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f == classIndex) {
+		if ((unsigned short)((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f == classIndex) {
 			return class;
 		} else {
 			class = *(Object **) ((unsigned char*) HEAP_REF(class, Object*)
@@ -1033,7 +1033,7 @@ static int16 newInstance(int32* sp, unsigned short classIndex) {
 #ifdef N_JAVA_LANG_CLASS_NEWINSTANCE
 int16 n_java_lang_Class_newInstance(int32* sp) {
 	Object *class = (Object *) (pointer) sp[0];
-	unsigned short classIndex = ((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
+	unsigned short classIndex = (unsigned short)((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
 
 	return newInstance(sp, classIndex);
 }
@@ -1053,7 +1053,7 @@ int16 n_java_lang_Class_getComponentType(int32 *sp) {
 	classIndex = getClassIndex(obj);
 	componentType = 0;
 	if (classIndex == (unsigned short) JAVA_LANG_CLASS_var) {
-		classIndex = ((java_lang_Class_c *)HEAP_REF(obj, Object*))-> cachedConstructor_f;
+		classIndex = (unsigned short)((java_lang_Class_c *)HEAP_REF(obj, Object*))-> cachedConstructor_f;
 		if (pgm_read_byte(&classes[classIndex].dimension) != 0) {
 			signed short componentTypeClasIndex;
 			componentTypeClasIndex = pgm_read_word(&classes[classIndex].dobjectSize);
@@ -1146,7 +1146,7 @@ extern Object* createArrayFromElementSize(unsigned short classIndex,
 int16 n_java_lang_reflect_Array_newArray(int32 *sp) {
 	Object* class = (Object*) (pointer) sp[0];
 	int32 size = sp[1];
-	unsigned short classIndex = ((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
+	unsigned short classIndex = (unsigned short)((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
 	Object* array = 0;
 	unsigned char elementSize = 0;
 
@@ -2398,7 +2398,7 @@ int16 n_java_lang_reflect_Method_invoke(int32 *sp) {
 #if defined(N_JAVA_LANG_CLASS_GETMETHOD)
 int16 n_java_lang_Class_getMethod(int32 *sp) {
 	Object *class = (Object*) (pointer) sp[0];
-	uint16 classIndex = ((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
+	uint16 classIndex = (uint16)((java_lang_Class_c *)HEAP_REF(class, Object*))-> cachedConstructor_f;
 	const char* className;
 
 	while (1) {
@@ -2474,7 +2474,7 @@ int16 n_java_lang_Class_getConstructor(int32 *sp) {
 	Object *cls = HEAP_REF((Object* )(pointer )sp[0], Object*);
 	Object *argsarray = HEAP_REF((Object* )(pointer )sp[1], Object*);
 
-	uint16 classIndex = ((java_lang_Class_c *)cls)-> cachedConstructor_f;
+	uint16 classIndex = (uint16)((java_lang_Class_c *)cls)-> cachedConstructor_f;
 	uint16 count = *((uint16*) argsarray + 1);
 
 	uint16 i;
